feat(EnvironmentsLists): add support for monitoring and updating grouped environment lists during drag-and-drop operations
